MAJOR COMMIT: Seperation of TEngine as it's own project.
This commit mainly seperates TEngine from this project into its own repo as a C++ library. Additionally we remove inclusion of raylib as TEngine comes pre packaged with raylib.

We also added small things here and there, populated the entity class a bit, added some methods to the labyrinth class and populated the functions of Game.cpp.
It's going really slow at the moment, but once everything comes together a bit for the first true execution, we should be kicking off fast!